Create sloving_using_reference.py
In here, I tried to create 3 reference lists for rows, columns and blocks (3X3 matrix) and using them as a reference the program will put a random value in (j, i)th position that is present in jth row, ith column and (j//3+(i//3)*3)th block and after putting the value in sudokuGrid the program will erase that valuse from jth row, ith column and (j//3+(i//3)*3)th block. But, there is a flaw.
As the program is chosing the values randomly there will be a higher possibility not to find out a number that is present in jth row, ith column and (j//3+(i//3)*3)th block.